[ISSUE-89] Refactor standard trait impls in core
