+++ core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "myopic-core"
-version = "1.1.2"
+version = "1.2.1"
 authors = ["Thomas Ball <tomoliverball@gmail.com>"]
 edition = "2018"
 license = "MIT"
@@ -18,3 +18,4 @@ itertools = "0.8"
 rand = "0.6.5"
 rand_pcg = "0.1.2"
 lazy_static = "1.2.0"
+anyhow = "1.0.35"
