+++ core/src/pieces/mod.rs
@@ -1,7 +1,7 @@
-use crate::bitboard::BitBoard;
-use crate::reflectable::Reflectable;
-use crate::Side;
-use crate::Square;
+use crate::{BitBoard, Side, Square};
+use anyhow::{anyhow, Error, Result};
+use std::fmt::{Display, Formatter};
+use std::str::FromStr;
 
 mod kings;
 mod knights;
@@ -16,19 +16,37 @@ pub enum Piece {
     WP, WN, WB, WR, WQ, WK,
     BP, BN, BB, BR, BQ, BK,
 }
+
+impl Display for Piece {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", format!("{:?}", self).to_lowercase())
+    }
+}
+
+impl FromStr for Piece {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let lower = s.to_lowercase();
+        Piece::all()
+            .find(|p| p.to_string() == lower)
+            .ok_or(anyhow!("Cannot parse {} as piece", s))
+    }
+}
+
 impl Piece {
     /// Create an iterator traversing over all pieces in order.
-    pub fn iter() -> impl Iterator<Item = Piece> {
+    pub fn all() -> impl Iterator<Item = Piece> {
         ALL.iter().cloned()
     }
 
     /// Create an iterator traversing over all white pieces in order.
-    pub fn iter_w() -> impl Iterator<Item = Piece> {
+    pub fn whites() -> impl Iterator<Item = Piece> {
         WHITE.iter().cloned()
     }
 
     /// Create an iterator traversing over all black pieces in order.
-    pub fn iter_b() -> impl Iterator<Item = Piece> {
+    pub fn blacks() -> impl Iterator<Item = Piece> {
         BLACK.iter().cloned()
     }
 
@@ -65,7 +83,7 @@ impl Piece {
     }
 
     /// Returns a slice containing all pieces belonging to the given side.
-    pub fn on_side(side: Side) -> impl Iterator<Item = Piece> {
+    pub fn of(side: Side) -> impl Iterator<Item = Piece> {
         match side {
             Side::White => (&WHITE).iter().cloned(),
             Side::Black => (&BLACK).iter().cloned(),
@@ -143,13 +161,6 @@ impl Piece {
     ];
 }
 
-/// We reflect a piece to it's correspondent on the opposite side.
-impl Reflectable for Piece {
-    fn reflect(&self) -> Self {
-        ALL[(*self as usize + 6) % 12]
-    }
-}
-
 /// Constant piece groupings.
 const ALL: [Piece; 12] = [
     Piece::WP,
@@ -183,3 +194,24 @@ const BLACK: [Piece; 6] = [
     Piece::BQ,
     Piece::BK,
 ];
+
+#[cfg(test)]
+mod test {
+    use crate::Piece;
+
+    #[test]
+    fn display() {
+        assert_eq!("wp", Piece::WP.to_string().as_str());
+        assert_eq!("br", Piece::BR.to_string().as_str());
+    }
+
+    #[test]
+    fn from_str() {
+        assert_eq!(Piece::WP, "wp".parse::<Piece>().unwrap());
+        assert_eq!(Piece::WP, "WP".parse::<Piece>().unwrap());
+        assert_eq!(Piece::BQ, "bq".parse::<Piece>().unwrap());
+        assert!("ba".parse::<Piece>().is_err());
+        assert!("bqs".parse::<Piece>().is_err());
+        assert!("wxk".parse::<Piece>().is_err());
+    }
+}
