+++ core/src/pieces/pawns.rs
@@ -1,6 +1,6 @@
 use crate::bitboard::BitBoard;
+use crate::square::Square;
 use crate::Side;
-use crate::Square;
 
 pub fn white_control(loc: Square, _whites: BitBoard, _blacks: BitBoard) -> BitBoard {
     WHITE_CONTROL[loc as usize]
@@ -181,7 +181,7 @@ const BLACK_CONTROL: [BitBoard; 64] = [
 #[cfg(test)]
 mod black_test {
     use crate::pieces::Piece;
-    use crate::Square::*;
+    use crate::square::Square::*;
 
     use super::*;
 
@@ -205,7 +205,7 @@ mod black_test {
 #[cfg(test)]
 mod white_test {
     use crate::pieces::Piece;
-    use crate::Square::*;
+    use crate::square::Square::*;
 
     use super::*;
